Move TestFeedExportBase to utils.
